Introduce SyncOMatic.Sync()
This method accepts a Diff as an input.
Missing Blobs and Trees are created in the target repositories according to each Diff definitions. A Commit that includes all the changes is created as well.

Optionally a Branch pointing at this commit can also be created. All created branches will bear the same name across repositories.